modifie template files

+++ build.sh
@@ -7,19 +7,46 @@
 GOTRACEBACK=all
 CURRENTDIR=`pwd`
 
-TEST_MODE=0
+TEST_MODE=0    #0:off, 1:after build, run test, 2:only run quickly
 AUTO_EXEC=0
 GODEP_MODE=1
 AUTO_GITCOMMIT=0
-HEROKU_MODE=0
-DOCKER_MODE=1  #0:off, 1:run server, 2:exec test on docker
+HEROKU_MODE=0  #0:off, 1:deploy server, 2:exec test on heroku
+DOCKER_MODE=0  #0:off, 1:run server,    2:exec test on docker
+
+GO_GET=0
+GO_LINT=0
+RESET_DB=0
+
+###########################################################
+# Run test quickly
+###########################################################
+if [ $TEST_MODE -eq 2 ]; then
+    go run ./cmd/ginserver/main.go
+    exit 0
+fi
+
+
+###########################################################
+# Reset Database (Restore)
+###########################################################
+if [ $RESET_DB -eq 1 ]; then
+    export DB_NAME=hiromaily
+    sh ./z_dbdata/setup.sh
+fi
+
 
 ###########################################################
 # Update all package
 ###########################################################
-#go get -u -v ./...
-#go get -d -v ./...
-#go get -u github.com/tools/godep
+if [ $GO_GET -eq 1 ]; then
+    go get -u -v ./...
+    #go get -d -v ./...
+
+    ### tools ###
+    go get -u github.com/tools/godep
+    go get -u github.com/lestrrat/go-server-starter/cmd/start_server
+fi
 
 
 ###########################################################
@@ -46,7 +73,9 @@ fi
 # go lint
 ###########################################################
 # it's too strict
-#golint ./...
+if [ $GO_LINT -eq 1 ]; then
+    golint ./...
+fi
 
 
 ###########################################################
@@ -59,8 +88,10 @@ fi
 # go build and install
 ###########################################################
 echo '============== go build -i -v -o; =============='
-rm -rf Godeps
-rm -rf ./vendor
+if [ $GODEP_MODE -eq 1 ]; then
+    rm -rf Godeps
+    rm -rf ./vendor
+fi
 
 #-n show just command for build
 #go build -i -n ./cmd/ginserver/
@@ -78,13 +109,20 @@ fi
 
 
 ###########################################################
+# cross-compile for linux
+###########################################################
+#GOOS=linux go install -v ./...
+
+
+###########################################################
 # test
 ###########################################################
 if [ $TEST_MODE -eq 1 ]; then
     echo '============== test =============='
 
     # Create Test Data
-    sh ./tests/setup.sh
+    export DB_NAME=hiromaily2
+    sh ./z_dbdata/setup.sh
 
     # Execute
     go test -v cmd/ginserver/*.go -f ../../configs/settings.toml
@@ -110,9 +148,9 @@ fi
 ###########################################################
 if [ $AUTO_EXEC -eq 1 ]; then
     echo '============== exec =============='
-    if [ $HEROKU_MODE -eq 1 ]; then
+    if [ $HEROKU_MODE -ne 0 ]; then
         #HEROKU ENV
-        export HEROKU_FLG=1
+        #export HEROKU_FLG=1
         #export CLEARDB_DATABASE_URL=mysql://be2ebea7cda583:49eef93c@us-cdbr-iron-east-04.cleardb.net/heroku_aa95a7f43af0311?reconnect=true
         #export REDIS_URL=redis://h:pf217irr4gts39d29o0012bghsi@ec2-50-19-83-130.compute-1.amazonaws.com:20819
 
@@ -123,10 +161,15 @@ if [ $AUTO_EXEC -eq 1 ]; then
     fi
 fi
 
-###########################################################
-# cross-compile for linux
-###########################################################
-#GOOS=linux go install -v ./...
+### Hot deplpy using go-server-starter
+# https://github.com/lestrrat/go-server-starter
+# http://takeshiyako.blogspot.jp/2015/10/go-lang-hot-deploy-with-go-server-starter.html
+
+# help
+#$GOPATH/bin/start_server --help
+
+
+
 
 
 ###########################################################
@@ -170,11 +213,20 @@ fi
 ###########################################################
 # heroku
 ###########################################################
-if [ $HEROKU_MODE -eq 1 ]; then
+if [ $HEROKU_MODE -gt 0 ]; then
     echo '============== heroku: git push =============='
     git push -f heroku master
+
+    if [ $HEROKU_MODE -eq 2 ]; then
+        # test
+        echo '============== heroku test =============='
+        #bash: go: command not found
+        #heroku run "go test -v /app/cmd/ginserver/*.go -f /app/configs/settings.toml" --app ginserver
+    fi
 fi
 
+
+###### e.g. command for heroku #####
 #heroku config:add HEROKU_FLG=1
 
 #heroku ps -a ginserver
